// frontend/src/company.d.ts
export interface CompanyIncomeStatement {
  accessNumber: string;
  symbol: string;
  year: number;
  quarter: number;
  form: string;
  startDate: string;
  endDate: string;
  report: {
    ic: {
      concept: string;
      label: string;
      unit: string;
      value: number;
    }[];
  };
}